Enemy Spawner (Read Desc)
Simple script to spawn enemies endlessly.

Script asks for number of enemy prefabs (2 in our case) then drag each to an element.

Then it asks for spawn interval (I set it to 5 in script but it is adjustable) which is time between each wave basically.

Lastly it asks for spawn points. I tested with just 1 element (just drag the empty object you want to the element and enemies will spawn there). I haven't been able to try using 2 due to my lag so please see if it works.

[NOTE]
This is quickly made and you can probably make it better optimized.